add cgzinfo tool
cgzinfo is a commandline tool that converts binary map files (*.cgz) to
a text format suitable for diff tools.

Copy the cgzinfo binary to your search path (for example /usr/local/bin)
and add a section
[diff "cgz"]
	textconv=cgzinfo -e -g -t
to .git/config.

The available options are:
 -e additionally print a list of entities
 -g additionally print the entire map geometry
 -t omit textures from the map geometry

With this filter, git diff and git blame are able to pinpoint specific
map changes - instead of just claiming "map files differ".

Move ".gitattributes" to ".git/info/attributes", if you don't want to
add .gitattributes to all branches.

No idea, if this works on windows.